Move assets of RaceDetail to client dir
