+++ app/views/races/show.html.haml
@@ -1,1 +1,1 @@
-= react_component 'RaceDetail', id: @id
+= react_component('RaceDetail', props: {id: @id}, prerender: false)
