+++ client/app/bundles/registration.jsx
@@ -1,4 +1,6 @@
 import ReactOnRails from 'react-on-rails';
+import RaceDetail from './components/RaceDetail';
 
-// This is how react_on_rails can see the HelloWorld in the browser.
-ReactOnRails.register({});
+ReactOnRails.register({
+  RaceDetail,
+});
